requirements.txt
PyQt5==5.15.10
pycryptodome==3.20.0
requests==2.31.0
pyinstaller==6.4.0
colorama==0.4.6
tqdm==4.66.2
pillow==10.2.0
psutil==5.9.8
pywin32==306
cryptography==42.0.2 